requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiokafka==0.12.0
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
bertopic==0.17.3
billiard==4.2.1
blis==1.3.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.22.0
confection==0.1.5
courlan==1.3.2
cryptography==45.0.7
cssselect==1.3.0
cymem==2.0.11
dateparser==1.2.2
DAWG-Python==0.7.2
DAWG2-Python==0.9.0
Django==5.2.5
django-emoji==2.2.2
docopt==0.6.2
emoji==2.14.0
environs==14.2.0
fake-useragent==2.2.0
fasttext==0.9.3
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.4
hdbscan==0.8.40
hf-xet==1.1.9
htmldate==1.9.3
huggingface-hub==0.34.4
idna==3.10
intervaltree==3.1.0
invoke==2.2.0
ipymarkup==0.9.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.5.2
jusText==3.0.2
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.4.0
lxml_html_clean==0.4.2
magic-filter==1.0.12
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==4.0.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
narwhals==2.5.0
natasha==1.6.0
navec==0.10.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.2
paramiko==4.0.0
pillow==11.0.0
playwright==1.55.0
playwright-stealth==1.0.6
plotly==6.3.0
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.32.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pybind11==3.0.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
PyNaCl==1.5.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
razdel==0.5.0
readability-lxml==0.8.4.1
redis==6.2.0
regex==2025.7.34
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
rsa==4.9.1
ru_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_lg-3.8.0/ru_core_news_lg-3.8.0-py3-none-any.whl#sha256=90bd584be86772b647e1b15cdc62e09d3a7182284ee23da95e19c2bf71aa0189
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==70.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
slovnet==0.6.0
smart_open==7.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
sqlparse==0.5.3
srsly==2.5.1
storage==0.0.4.3
sympy==1.13.1
Telethon==1.40.0
thinc==8.3.6
threadpoolctl==3.6.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.22.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.56.0
triton==3.1.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.9.post2
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.17.3
yargy==0.16.0
yarl==1.20.1
